requirements.txt
streamlit~=1.44.1
fastapi~=0.115.12
uvicorn
httpx
requests~=2.32.3
pydantic~=2.11.3
openai~=1.76.2
python-dotenv~=1.1.0
face_recognition~=1.3.0
Authlib~=1.5.2
numpy~=2.2.5
pdf2image~=1.17.0
pillow~=11.2.1
face_recognition_models~=0.3.0
requests-oauthlib~=2.0.0